feat: add tool.gpt and changes other scripts to make this an obot tool.
